% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultaDespesa.R
\name{consultaDespesa_body}
\alias{consultaDespesa_body}
\title{consultaDespesa Body}
\usage{
consultaDespesa_body(ano, codigoOrgao = "Consolidado", codigoUo = "",
  codigoUnidadeGestora = "", codigoFonteRecursos = "",
  codigoTipoLicitacao = "", codigoFuncao = "", codigoSubfuncao = "",
  codigoPrograma = "", codigoAcao = "", codigoFuncionalProgramatica = "",
  codigoMunicipio = "", codigoCategoria = "", codigoGrupo = "",
  codigoModalidade = "", codigoElemento = "", naturezaDespesa = "",
  flagCredor = 0, cgcCpfCredor = "", nomeCredor = "", flagEmpenhado = 1,
  flagLiquidado = 0, flagPago = 0)
}
\arguments{
\item{ano}{ano}

\item{codigoOrgao}{codigoOrgao}
}
\description{
Generates the BODY for the soap request for
consultaDespesa
}

