% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultaDespesa.R
\name{consultaDespesa_get}
\alias{consultaDespesa_get}
\title{consultaDespesa Get}
\usage{
consultaDespesa_get(ano, codigoOrgao = "Consolidado", ...)
}
\arguments{
\item{ano}{ano}

\item{codigoOrgao}{codigoOrgao}
}
\description{
Get data from the webservice.
}
\examples{
\dontrun{
consulta <- consultaDespesa_get(ano = 2016, codigoOrgao = "42000")
# You can save the result to a xml file using:
consultaDespesa_get(ano = 2016, codigoOrgao = "42000", httr::write_disk("file.xml"))
}

}

