% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultaDespesa.R
\name{consultaDespesa_parse}
\alias{consultaDespesa_parse}
\title{consultaDespesa Parse}
\usage{
consultaDespesa_parse(xml)
}
\arguments{
\item{xml}{a xml document (a path to a xml file or a connection returned by
consultaDespesa or a file read by read_xml function)}
}
\description{
Parse results from consultaDespesa.
}

