% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultaDespesa.R
\name{get_child}
\alias{get_child}
\title{get_child}
\usage{
get_child(itens, child)
}
\arguments{
\item{itens}{itens xml_nodeset.}

\item{child}{name of the field to get}
}
\description{
An util that gets the child of all itens.
}

