% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultaDespesa.R
\name{consultaDespesa_body}
\alias{consultaDespesa_body}
\title{consultaDespesa Body}
\usage{
consultaDespesa_body(ano, codigoOrgao = "Consolidado", codigoUo = "",
  codigoUnidadeGestora = "", codigoFonteRecursos = "",
  codigoTipoLicitacao = "", codigoFuncao = "", codigoSubfuncao = "",
  codigoPrograma = "", codigoAcao = "", codigoFuncionalProgramatica = "",
  codigoMunicipio = "", codigoCategoria = "", codigoGrupo = "",
  codigoModalidade = "", codigoElemento = "", naturezaDespesa = "",
  flagCredor = 0, cgcCpfCredor = "", nomeCredor = "", flagEmpenhado = 1,
  flagLiquidado = 0, flagPago = 0)
}
\arguments{
\item{ano}{ano}

\item{codigoOrgao}{codigoOrgao}

\item{codigoUo}{default: ""}

\item{codigoUnidadeGestora}{default: ""}

\item{codigoFonteRecursos}{default: ""}

\item{codigoTipoLicitacao}{default: ""}

\item{codigoFuncao}{default: ""}

\item{codigoSubfuncao}{default: ""}

\item{codigoPrograma}{default: ""}

\item{codigoAcao}{default: ""}

\item{codigoFuncionalProgramatica}{default: ""}

\item{codigoMunicipio}{default: ""}

\item{codigoCategoria}{default: ""}

\item{codigoGrupo}{default: ""}

\item{codigoModalidade}{default: ""}

\item{codigoElemento}{default: ""}

\item{naturezaDespesa}{default: ""}

\item{flagCredor}{default: 0}

\item{cgcCpfCredor}{default: ""}

\item{nomeCredor}{default: ""}

\item{flagEmpenhado}{default: 1}
}
\description{
Generates the BODY for the soap request for
consultaDespesa
}

