% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultaDespesa.R
\name{consultaDespesa_get}
\alias{consultaDespesa_get}
\title{consultaDespesa Get}
\usage{
consultaDespesa_get(ano, codigoOrgao = "Consolidado", ..., codigoUo = "",
  codigoUnidadeGestora = "", codigoFonteRecursos = "",
  codigoTipoLicitacao = "", codigoFuncao = "", codigoSubfuncao = "",
  codigoPrograma = "", codigoAcao = "", codigoFuncionalProgramatica = "",
  codigoMunicipio = "", codigoCategoria = "", codigoGrupo = "",
  codigoModalidade = "", codigoElemento = "", naturezaDespesa = "",
  flagCredor = 0, cgcCpfCredor = "", nomeCredor = "", flagEmpenhado = 1,
  flagLiquidado = 0, flagPago = 0)
}
\arguments{
\item{ano}{ano}

\item{codigoOrgao}{codigoOrgao}

\item{...}{other arguments passed to httr::POST.}

\item{codigoUo}{default: ""}

\item{codigoUnidadeGestora}{default: ""}

\item{codigoFonteRecursos}{default: ""}

\item{codigoTipoLicitacao}{default: ""}

\item{codigoFuncao}{default: ""}

\item{codigoSubfuncao}{default: ""}

\item{codigoPrograma}{default: ""}

\item{codigoAcao}{default: ""}

\item{codigoFuncionalProgramatica}{default: ""}

\item{codigoMunicipio}{default: ""}

\item{codigoCategoria}{default: ""}

\item{codigoGrupo}{default: ""}

\item{codigoModalidade}{default: ""}

\item{codigoElemento}{default: ""}

\item{naturezaDespesa}{default: ""}

\item{flagCredor}{default: 0}

\item{cgcCpfCredor}{default: ""}

\item{nomeCredor}{default: ""}

\item{flagEmpenhado}{default: 1}

\item{flagLiquidado}{default: 0}

\item{flagPago}{default: 0}
}
\description{
Get data from the webservice.
}
\examples{
\dontrun{
consulta <- consultaDespesa_get(ano = 2016, codigoOrgao = "42000")
# You can save the result to a xml file using:
consultaDespesa_get(ano = 2016, codigoOrgao = "42000", httr::write_disk("file.xml"))
}

}

